pyproject.toml
[project]
name = "skill-based-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ale-py>=0.11.2",
    "atariari",
    "gymnasium[atari]>=1.2.3",
    "huggingface-hub>=0.36.0",
    "huggingface-sb3>=3.0",
    "imageio>=2.37.2",
    "ipykernel>=7.1.0",
    "kornia>=0.8.2",
    "moviepy>=2.2.1",
    "opencv-python>=4.11.0.86",
    "pyyaml>=6.0.3",
    "rl-zoo3>=2.7.0",
    "scikit-learn>=1.8.0",
    "shimmy>=0.2.1",
    "stable-baselines3[extra]>=2.7.1",
    "tensorflow>=2.20.0",
    "torchvision>=0.24.1",
    "wandb>=0.23.1",
]

[tool.uv.sources]
atariari = { git = "https://github.com/Sopralapanca/atari-representation-learning.git" }
